fs: handle both CONFIG_OVERLAY(|FS)_FS option
Before overlayfs was put in the kernel, the option name was
`CONFIG_OVERLAYFS_FS` and the module name was `overlayfs`. After its
integration, the option name is `CONFIG_OVERLAY_FS` and the module name
is `overlay`.

The option name was handled in 2da20859ccb5 but the module name was not
handled correctly. Closes #2.