Send console output to correct serial device
Qemu is configured to treat ttyS1 as a serial device, so this patch
instructs the kernel to use the correct device and disables the quiet
option so that users can see the console output.

Further options could be added, such as 'debug' and 'ignore_loglevel'
for increased debugging, but this should provide a basic working serial
console.